Fixed patient apis issues

+++ backend/src/main/java/com/ilouse/medicalhub/repo/PatientRepo.java
@@ -5,5 +5,5 @@ import com.ilouse.medicalhub.model.Patient;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.*;
 
-public interface PatientRepo extends JpaRepository<Patient, Integer> {
+public interface PatientRepo extends JpaRepository<Patient, Long> {
 }
